requirements.txt
mcp==1.26.0
spotipy==2.26.0
starlette==0.46.1
uvicorn==0.34.0
python-dotenv==1.0.1
